% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{get_progression_times}
\alias{get_progression_times}
\title{Sample date-time values from fire progression polygons}
\usage{
get_progression_times(
  prog,
  pts,
  prog_datetime,
  formats = c("ymdHMS", "ymdHM", "ymd"),
  prog_tz = "Australia/NSW"
)
}
\arguments{
\item{prog}{An \code{sf} spatial data frame of fire progression polygons. The
data must have a coordinate reference system defined and the progression
polygons must not overlap (both of these conditions will be checked).}

\item{pts}{An \code{sf} spatial data frame of sample points as generated by
function \code{make_sample_points()}. It is assumed that each sample point
has attributes to relate it to a back-burning line.}

\item{prog_datetime}{The name of the column in the input \code{progressions} data
frame with date-time values.}

\item{formats}{WRITE ME!!!}

\item{prog_tz}{WRITE ME!!!}
}
\value{
A copy of the input \code{sample_points} spatial data frame with an
  additional column \code{'prog_datetime'} containing the date-time values.
}
\description{
Sample date-time values from fire progression polygons
}
\examples{
\dontrun{
library(sf)
library(backburning)

# Load a layer of back-burning lines and generate sample points at the
# desired spacing
dat_bb <- st_read("path/to/bb_lines")
dat_pts <- make_sample_points(dat_bb, bb_id = "OID",
                              line_spacing = 1000,
                              line_half_length = 5000,
                              point_spacing = 500)

# Load a corresponding layer of fire progression polygons and get
# the date-time value for each sample point
dat_prog <- st_read("path/to/progressions")

dat_pts <- get_progression_times(prog, pts, prog_datetime = "date")
}

}
