% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{make_sampling_lines}
\alias{make_sampling_lines}
\title{Generate sampling lines along one or more back-burning line features}
\usage{
make_sampling_lines(
  bb_lines,
  bb_id = "OID",
  step_length = 500,
  sample_length = 5000,
  stop_if_returning = TRUE,
  smoothing_bw = 1000
)
}
\arguments{
\item{bb_lines}{An \code{sf} spatial data frame containing one or more line
features representing back-burning lines. The data must have a coordinate
reference system defined with metres as map units (e.g. NSW Lambert/GDA94
EPSG:3308).}

\item{bb_id}{(character; default "OID") The name of a column in the
\code{bb_lines} data frame that uniquely identifies back-burning line
features.}

\item{step_length}{A single numeric value for the distance in metres
between sampling lines along each back-burning line feature.}

\item{sample_length}{A single numeric value specifying the maximum distance
in metres that a sample line will extend either side of a back-burning
line.}

\item{stop_if_returning}{(logical) If \code{TRUE} (default), each sample line
must be moving away from its associated back-burning line along its whole
length. This is intended to avoid cases where a sample line approaches or
crosses another section of a wiggly back-burning line.}

\item{smoothing_bw}{A single numeric value for the bandwidth (metres) of the
Gaussian kernel filter used to smooth each back-burning feature. The
default value of 1000m seems to give good results.}
}
\value{
An \code{sf} spatial data frame containing the sampling lines.
}
\description{
Generate sampling lines along one or more back-burning line features
}
