% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{make_sample_points}
\alias{make_sample_points}
\title{Generate lines of sample points across back-burning line features}
\usage{
make_sample_points(
  bb_lines,
  bb_id = "OID",
  line_spacing,
  line_half_length,
  point_spacing,
  increasing_distance = c("all", "parent", "none"),
  smoothing_bw = 1000
)
}
\arguments{
\item{bb_lines}{An \code{sf} spatial data frame containing one or more line
features representing back-burning lines. The data must have a coordinate
reference system defined with metres as map units (e.g. NSW Lambert/GDA94
EPSG:3308).}

\item{bb_id}{(character; default "OID") The name of a column in the
\code{bb_lines} data frame that uniquely identifies back-burning line
features.}

\item{line_spacing}{A single numeric value for the distance in metres
between sampling lines along each back-burning line feature.}

\item{line_half_length}{A single numeric value specifying the maximum
distance in metres that a sampling line will extend on either side of the
back-burning line.}

\item{point_spacing}{A single numeric value specifying the uniform distance
between sample points along each sampling line. The first point is always
positioned at the intersection of the sampling line and the reference
back-burning line.}

\item{increasing_distance}{(character) One of \code{('all', 'parent',
'none')}. May be abbreviated. If \code{'all'} (default), each sample point
must lie at a greater distance to all back-burning lines than the preceding
point AND must be closer to the parent back-burning line than any other; if
\code{'parent'}, a similar increasing distance rule is applied but only in
relation to the parent back-burning line that the sample points are
associated with; if \code{'none'}, no distance rule applies.}

\item{smoothing_bw}{A single numeric value for the bandwidth (metres) of the
Gaussian kernel filter used to smooth each back-burning feature. The
default value of 1000m seems to give good results.}
}
\value{
An \code{sf} spatial data frame of sample point features with the
  following columns:

  \code{refid} Identifier of the input back-burning line feature that a point
  is sampling, with values taken from the input data column specified by the
  \code{bb_id} argument.

  \code{line_index} Integer index (from 1) of the perpendicular sample line
  on which the point was positioned.

  \code{segment} One of \code{'X'} (point at intersection with back-burning
  line); \code{'L'} (point on left of line); \code{'R'} (point on right of
  line). Note that left and right are relative to the digitizing direction
  (i.e. order of vertices) of the back-burning line.

  \code{dist_line} Distance of the point along the sample line segment on
  which it was positioned, relative to the parent back-burning line.

  \code{dist_parent} Shortest distance from the point to the parent
  back-burning line. This can be less than \code{dist_line} if the parent
  line is wiggly.

  \code{dist_all} Shortest distance from the point to \strong{any}
  back-burning line. This column will only be present if the argument
  \code{increasing_distance == 'all'}.

  \code{geom} Point geometry, projected in the same coordinate reference
  system as the input line features.
}
\description{
Given a set of input line features representing back-burning lines, this
function places sample lines at regular intervals along each input feature,
perpendicular to the local feature angle, and then generates uniformly spaced
points along each sample line. Local angles are determined from a smoothed
version of the back-burning line to minimize the influence of any local kinks
and turns. Sample lines are located at, and either side of, the mid-point of
each input feature. At each sampling line, a sample point is located where
the line intersects the back-burning line, then further points are placed
along the line at uniform spacing. In cases where a back-burning line is
curved or convoluted it is possible for sampling lines from one section to
approach or cross other sections. Setting the argument
\code{increasing_distance} to \code{TRUE} (the default) will test for such
cases and prune sampling lines as required. This can result in different
numbers of sample points per line being returned.
}
\details{
This function calls the non-exported helper function
\code{make_sample_lines()} to generate the sample lines along which the
sample points are placed.
}
\examples{
\dontrun{
libary(backburning)
library(sf)

# Load back-burning line features from a GeoPackage layer, shapefile etc.
dat_bb <- st_read(...)

# Generate points on sampling lines placed at 1km intervals along each
# back-burning line. Sampling lines extend 5km either side of the back-burning
# line and points are placed every 500m.
#
dat_sample_points <- make_sample_points(bb,
                                        line_spacing = 1000,
                                        line_half_length = 5000,
                                        point_spacing = 500)
}

}
